separate tree-util to module cleanup
