+++ homework-01/package.json
@@ -3,9 +3,6 @@
   "version": "1.0.0",
   "description": "otus tree homework",
   "main": "server.js",
-  "scripts": {
-    "start": "node server.js"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ilnoor/gabdushev-ilnur-otus.git"
